requirements.txt
alabaster==0.7.11
asn1crypto==0.24.0
astroid==2.0.2
autopep8==1.3.5
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.6.1
bleach==2.1.3
bs4==0.0.1
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.5.3
cryptography==2.3.1
cycler==0.10.0
decorator==4.3.0
docutils==0.14
EasyProcess==0.2.3
entrypoints==0.2.3
html5lib==1.0.1
idna==2.7
imagesize==1.0.0
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
isort==4.3.4
jedi==0.12.1
jeepney==0.3.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.0
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.0
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.4
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.15.0
numpydoc==0.8.0
packaging==17.1
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
pkg-resources==0.0.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
psutil==5.4.7
ptyprocess==0.6.0
pycodestyle==2.4.0
pycparser==2.18
pyflakes==2.0.0
Pygments==2.2.0
pylint==2.1.1
pyparsing==2.2.0
PyQt5==5.9.2
python-dateutil==2.7.3
pytz==2018.5
PyVirtualDisplay==0.2.1
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.0
QtPy==1.4.2
requests==2.19.1
rope==0.11.0
scipy==1.1.0
seaborn==0.9.0
SecretStorage==3.0.1
selenium==3.14.0
Send2Trash==1.5.0
simplegeneric==0.8.1
sip==4.19.8
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.6
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
typed-ast==1.1.0
typing==3.6.4
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.0
wrapt==1.10.11
